style: split artifacts mod and create contract and bytecode mods
