+++ ethers-solc/src/lib.rs
@@ -32,11 +32,12 @@ pub mod utils;
 pub use filter::{FileFilter, TestFileFilter};
 
 use crate::{
-    artifacts::{Contract, Sources},
+    artifacts::Sources,
     cache::SolFilesCache,
     contracts::VersionedContracts,
     error::{SolcError, SolcIoError},
 };
+use artifacts::contract::Contract;
 use error::Result;
 use semver::Version;
 use std::path::{Path, PathBuf};
